+++ components/bridge/dao/impl/src/main/java/com/flipperdevices/bridge/dao/impl/api/parsers/url/FFFUrlEncoder.kt
@@ -10,18 +10,23 @@ class FFFUrlEncoder : LogTagProvider {
     override val TAG = "FFFUrlEncoder"
 
     fun keyToUri(path: FlipperFilePath, fff: FlipperFileFormat): URL {
-        val query = listOf(QUERY_KEY_PATH to path.pathToKey).plus(fff.orderedDict)
-            .filterNot { it.first.isBlank() || it.second.isBlank() }
-            .joinToString(QUERY_DELIMITED_CHAR) {
-                val field = URLEncoder.encode(it.first.trim(), "UTF-8")
-                val value = URLEncoder.encode(it.second.trim(), "UTF-8")
-                    .replace("%2F", "/") // We want safe / for readability
-                "$field$QUERY_VALUE_DELIMITED_CHAR$value"
-            }
+        val params = listOf(QUERY_KEY_PATH to path.pathToKey).plus(fff.orderedDict)
+        val query = encodeQuery(params)
         return URL(
             PREFFERED_SCHEME,
             PREFFERED_HOST,
             "$PATH_FOR_FFF_LIHK#$query"
         )
+    }
+
+    fun encodeQuery(params: List<Pair<String, String>>): String {
+        return params
+            .filterNot { it.first.isBlank() || it.second.isBlank() }
+            .joinToString(QUERY_DELIMITED_CHAR) {
+                val field = URLEncoder.encode(it.first.trim(), QUERY_VALUE_CHARSET)
+                val value = URLEncoder.encode(it.second.trim(), QUERY_VALUE_CHARSET)
+                    .replace("%2F", "/") // We want safe / for readability
+                "$field$QUERY_VALUE_DELIMITED_CHAR$value"
+            }
     }
 }