+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 1.3.1 - In progress
 
 - [Feature] Add forbidden frequency dialog
+- [Feature] Rework all system dialog to custom
+- [Feature] New share flow
 
 # 1.3.0